[REV-289] refactor: 리뷰 수신자 전체 조회 로직 수정

+++ src/main/java/com/devcourse/ReviewRanger/participation/dto/response/ReceiverResponse.java
@@ -11,6 +11,6 @@ public record ReceiverResponse(
 	String receiverName,
 
 	@Schema(description = "응답자 수")
-	Integer responserCount
+	Long responserCount
 ) {
 }
